updated custom query (#90)
* updated custom query

updated the syntax of the custom query file to match versions 13 and 14 of postgres

* Updated config

Updated the config with notes for managed database environments, specifically from aws and azure.

* updated config

commented out the SSL Location parameters since they're set to linux only destinations. As it is, the template file would fail in a windows environment. Also set ENABLE_SSL to false since that's the documented default.

* Added custom query to build

Added the custom query file to the build file so it will download on install.